create shared getActiveSession selector and remove duplicates
