+++ lib/interviewer/hooks/useExternalData.js
@@ -9,7 +9,7 @@ import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { getVariableTypeReplacements } from '../containers/withExternalData';
 import { getAssetManifest, getProtocolCodebook } from '../selectors/protocol';
-import { getActiveSession } from '../selectors/session';
+import { getActiveSession } from '../selectors/shared';
 import getParentKeyByNameValue from '../utils/getParentKeyByNameValue';
 import loadExternalData from '../utils/loadExternalData';
