+++ lib/interviewer/selectors/protocol.js
@@ -2,6 +2,7 @@ import { createSelector } from '@reduxjs/toolkit';
 import { get } from 'es-toolkit/compat';
 import { v4 as uuid } from 'uuid';
 import { getStageSubject } from './prop';
+import { getActiveSession } from './shared';
 
 const DefaultFinishStage = {
   // `id` is used as component key; must be unique from user input
@@ -9,9 +10,6 @@ const DefaultFinishStage = {
   type: 'FinishSession',
   label: 'Finish Interview',
 };
-
-const getActiveSession = (state) =>
-  state.activeSessionId && state.sessions[state.activeSessionId];
 
 const getInstalledProtocols = (state) => state.installedProtocols;
