+++ lib/interviewer/selectors/session.ts
@@ -1,19 +1,7 @@
 import type { Stage } from '@codaco/shared-consts';
-import { getProtocolStages } from './protocol';
 import { createSelector } from '@reduxjs/toolkit';
-import type { RootState } from '../store';
-
-const getActiveSessionId = (state: RootState) => state.activeSessionId;
-
-const getSessions = (state: RootState) => state.sessions;
-
-export const getActiveSession = createSelector(
-  getActiveSessionId,
-  getSessions,
-  (activeSessionId, sessions) => {
-    return sessions[activeSessionId]!;
-  },
-);
+import { getProtocolStages } from './protocol';
+import { getActiveSession } from './shared';
 
 export const getStageIndex = createSelector(getActiveSession, (session) => {
   return session.currentStep;
